[update] - split views apart
