+++ backend/api/fixtures/tools_fixture.json
@@ -6,7 +6,9 @@
             "name": "Get Location Weather",
             "description": "A tool to get the current weather information for a single location.",
             "script": "import python_weather\n\nasync def get_weather_information(single_location: str) -> str:\n    \"\"\"Get the current weather of the specified single location.\n\n    Args:\n        location (str): The single location to get the current weather for.\n\n    Returns:\n        str: A string containg various metrics about the current weather for that single location.\n    \"\"\"\n    async with python_weather.Client(unit=python_weather.IMPERIAL) as client:\n        # Get the weather for the location\n        weather = await client.get(single_location)\n        output = \"\"\n        output += f\"Datetime: {weather.datetime}\"\n        output += f\"Description: {weather.description}\"\n        output += f\"Feels Like: {weather.feels_like}\"\n        output += f\"Humidity: {weather.humidity}\"\n        output += f\"Kind: {weather.kind}\"\n        output += f\"Precipitation: {weather.precipitation}\"\n        output += f\"Temperature: {weather.temperature}\"\n        output += f\"Ultraviolet: {weather.ultraviolet}\"\n        output += f\"Visibility: {weather.visibility}\"\n        output += f\"Wind Direction: {weather.wind_direction}\"\n        output += f\"Wind Speed: {weather.wind_speed}\"\n        return output",
-            "user": 1
+            "user": 1,
+            "created_at": "2025-01-20T16:00:21.753646",
+            "updated_at": "2025-01-20T16:00:21.753646"
         }
     }
 ]
