+++ backend/api/views/conversation.py
@@ -1,0 +1,72 @@
+import pytz
+import datetime
+from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.response import Response
+
+from ..serializers.conversation_serializer import ConversationSerializer
+from ..serializers.conversation_detail_serializer import ConversationDetailSerializer
+
+from ..models.conversation import Conversation
+
+class ConversationListCreateView(ListCreateAPIView):
+    queryset = None
+    serializer_class = ConversationSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return Conversation.objects.filter(user=self.request.user)
+
+    def get(self, request):
+        conversations = self.get_queryset()
+        serializer = self.get_serializer(conversations, many=True)
+
+        utc = pytz.UTC  # UTC timezone instance
+
+        for conversation in serializer.data:
+            # Parse and make the `created_at` aware of UTC
+            created_at = datetime.datetime.fromisoformat(
+                conversation["created_at"].replace("Z", "+00:00")
+            ).astimezone(utc)
+
+            # Get current time in UTC
+            now_utc = datetime.datetime.now(datetime.UTC)
+
+            # Calculate the time difference in days
+            days_difference = (now_utc.date() - created_at.date()).days
+
+            if days_difference == 0:
+                conversation["grouping"] = "Today"
+            elif days_difference <= 7:
+                conversation["grouping"] = "This Week"
+            elif now_utc.month == created_at.month:
+                conversation["grouping"] = "This Month"
+            else:
+                conversation["grouping"] = "Old"
+
+        return Response(serializer.data)
+
+    def perform_create(self, serializer):
+        serializer.save(user=self.request.user)
+
+
+class ConversationDetailView(RetrieveUpdateDestroyAPIView):
+    queryset = None
+    serializer_class = ConversationDetailSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        return Conversation.objects.filter(user=self.request.user)
+
+    def get_object(self):
+        conversation = super().get_object()
+
+        if conversation.user != self.request.user:
+            raise PermissionDenied(
+                "You do not have permission to access this conversation."
+            )
+
+        return conversation
+
+    def perform_update(self, serializer):
+        serializer.save(user=self.request.user)
